pyproject.toml
[project]
name = "gi-talker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "discord-py>=2.6.4",
    "numpy>=1.23,<2.0",
    "pynacl>=1.6.0",
    "python-dotenv>=1.1.1",
    "soundfile>=0.13.1",
]

[project.optional-dependencies]
melotts = [
    "torch>=2.4.0",
    "torchaudio>=2.4.0",
    "melotts @ git+https://github.com/myshell-ai/MeloTTS.git",
    "cached_path>=1.5.2",
    "transformers==4.27.4",
    "num2words==0.5.12",
    "unidic-lite==1.0.8",
    "unidic==1.1.0",
    "mecab-python3==1.0.9",
    "pykakasi==2.2.1",
    "fugashi==1.3.0",
    "g2p_en==2.1.0",
    "anyascii==0.3.2",
    "jamo==0.4.1",
    "gruut[de,es,fr]==2.2.3",
    "g2pkk>=0.1.1",
    "librosa==0.9.1",
    "pydub==0.25.1",
    "eng_to_ipa==0.0.2",
    "inflect==7.0.0",
    "Unidecode==1.3.7",
    "pypinyin==0.50.0",
    "cn2an==0.5.22",
    "jieba==0.42.1",
    "gradio>=4.0.0",
    "langid==1.1.6",
    "tqdm>=4.66.0",
    "tensorboard==2.16.2",
    "loguru==0.7.2",
]

[tool.uv]
package = true

[tool.uv.sources]
"gi-talker" = { path = "src" }
